Fix review permissions problem

+++ app/controllers/reviews_controller.rb
@@ -1,11 +1,9 @@
 class ReviewsController < ApplicationController
+  before_filter :authenticate_user!
 
   expose(:review)
   expose(:product)
   expose(:category)
-
-  def edit
-  end
 
   def create
     self.review = Review.new(review_params)
@@ -13,15 +11,19 @@ class ReviewsController < ApplicationController
 
     if review.save
       product.reviews << review
-      redirect_to category_product_url(product.category, product), notice: 'Review was successfully created.'
+      redirect_to [product.category, product], notice: 'Review was successfully created.'
     else
       render action: 'new'
     end
   end
 
   def destroy
+    unless current_user == review.user or current_user.try(:admin?)
+      redirect_to [product.category, product], notice: 'You are not allowed to remove this review.'
+    end
+
     review.destroy
-    redirect_to category_product_url(product.category, product), notice: 'Review was successfully destroyed.'
+    redirect_to [product.category, product], notice: 'Review was successfully destroyed.'
   end
 
   private
